+++ src/main/java/com/matchvagas/backend/repository/CidadeRepository.java
@@ -1,14 +1,12 @@
 package com.matchvagas.backend.repository;
 
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
 import com.matchvagas.backend.entity.Cidade;
 import com.matchvagas.backend.entity.Estado;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import java.util.List;
 
 @Repository
-public interface CidadeRepository extends JpaRepository<Cidade, Integer> {
+public interface CidadeRepository extends JpaRepository<Cidade, Long> {
     List<Cidade> findByEstado(Estado estado);
 }
