+++ src/main/java/com/matchvagas/backend/repository/DepartamentosRepository.java
@@ -1,11 +1,8 @@
 package com.matchvagas.backend.repository;
 
+import com.matchvagas.backend.entity.Departamentos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.matchvagas.backend.entity.Departamentos;
-
 @Repository
-public interface DepartamentosRepository extends JpaRepository<Departamentos, Integer> {
-    
-}
+public interface DepartamentosRepository extends JpaRepository<Departamentos, Long> {}
