+++ src/main/java/com/matchvagas/backend/repository/ModalidadeRepository.java
@@ -1,11 +1,8 @@
 package com.matchvagas.backend.repository;
 
+import com.matchvagas.backend.entity.Modalidade;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.matchvagas.backend.entity.Modalidade;
-
 @Repository
-public interface ModalidadeRepository extends JpaRepository<Modalidade, Integer> {
-    
-}
+public interface ModalidadeRepository extends JpaRepository<Modalidade, Long> {}
