+++ src/main/java/com/matchvagas/backend/repository/StatusVagaRepository.java
@@ -1,13 +1,11 @@
 package com.matchvagas.backend.repository;
 
+import com.matchvagas.backend.entity.StatusVaga;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import com.matchvagas.backend.entity.StatusVaga;
-
 import java.util.Optional;
 
 @Repository
-public interface StatusVagaRepository extends JpaRepository<StatusVaga, Integer> {
+public interface StatusVagaRepository extends JpaRepository<StatusVaga, Long> {
     Optional<StatusVaga> findByDescricao(String descricao);
-} 
+}
